+++ ShadowRunHelper.Win/UI/Parts/CategoryEntry.xaml
@@ -2,7 +2,7 @@
     x:Class="ShadowRunHelper.Win.UI.CategoryEntry"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:ShadowRunHelper.Win.UI.Parts"
+    xmlns:local="using:ShadowRunHelper.Win.UI"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:CharModel="using:ShadowRunHelper.CharModel"
