+++ ShadowRunHelper.Win/UI/Parts/CategoryEntry.xaml.cs
@@ -1,180 +1,178 @@
-﻿using System;
+﻿using ShadowRunHelper.CharModel;
+using ShadowRunHelper.Model;
+using ShadowRunHelper.UI;
+using ShadowRunHelper.UI.Edit;
+using System;
 using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
-using Windows.UI.Xaml.Navigation;
-
-// Die Elementvorlage "Benutzersteuerelement" wird unter https://go.microsoft.com/fwlink/?LinkId=234236 dokumentiert.
 
 namespace ShadowRunHelper.Win.UI
 {
     public sealed partial class CategoryEntry : UserControl
     {
+        #region Variables
+        readonly AppModel Model = AppModel.Instance;
+        Thing CurrentThing => (DataContext as Thing);
+        #endregion
         public CategoryEntry()
         {
             this.InitializeComponent();
-            EntryItem.ContentTemplate = ItemItem;
+            GetStuff();
+            DefaultTemplate();
+        }
+        DataTemplate Default;
+        DataTemplate Expanded;
+        private void GetStuff()
+        {
+            Default = (DataTemplate)Resources[TypeHelper.ThingDefToString(CurrentThing.ThingType, false) + "Item"];
+            Expanded = (DataTemplate)Resources[TypeHelper.ThingDefToString(CurrentThing.ThingType, false) + "ItemX"];
         }
 
-        public void Expand()
+        public void ExpandedTemplate()
         {
-            EntryItem.ContentTemplate = ItemItemX;
-
-        }
-
-        internal void Shrink()
-        {
-            EntryItem.ContentTemplate = ItemItem;
-
-        }
-        public void ChangeTemplate()
-        {
-            switch ((int)ThingDefs.Handlung)
+            EntryItem.ContentTemplate = Expanded;
+            if (CurrentThing.ThingType == ThingDefs.Handlung)
             {
-                case (int)ThingDefs.Handlung:
-                //                    NewTemplate = HandlungItem;
-                //                    NewTemplateX = HandlungItemX;
-                //                    if (!SettingsModel.I.TutorialHandlungShown)
-                //                    {
-                //#pragma warning disable CS4014
-                //                        new Tutorial(30, 31).ShowAsync();
-                //#pragma warning restore CS4014
-                //                        SettingsModel.I.TutorialHandlungShown = true;
-                //                    }
-                //                    break;
-                //                case (int)ThingDefs.Fertigkeit:
-                //                    NewTemplate = FertigkeitItem;
-                //                    NewTemplateX = FertigkeitItemX;
-                //                    break;
-                //                case (int)ThingDefs.Item:
-                //                    NewTemplate = ItemItem;
-                //                    NewTemplateX = ItemItemX;
-                //                    break;
-                //                case (int)ThingDefs.Programm:
-                //                    NewTemplate = ProgrammItem;
-                //                    NewTemplateX = ProgrammItemX;
-                //                    break;
-                //                case (int)ThingDefs.Munition:
-                //                    NewTemplate = MunitionItem;
-                //                    NewTemplateX = MunitionItemX;
-                //                    break;
-                //                case (int)ThingDefs.Implantat:
-                //                    NewTemplate = ImplantatItem;
-                //                    NewTemplateX = ImplantatItemX;
-                //                    break;
-                //                case (int)ThingDefs.Vorteil:
-                //                    NewTemplate = EigenschaftItem;
-                //                    NewTemplateX = EigenschaftItemX;
-                //                    break;
-                //                case (int)ThingDefs.Nachteil:
-                //                    NewTemplate = EigenschaftItem;
-                //                    NewTemplateX = EigenschaftItemX;
-                //                    break;
-                //                case (int)ThingDefs.Connection:
-                //                    NewTemplate = ConnectionItem;
-                //                    NewTemplateX = ConnectionItemX;
-                //                    break;
-                //                case (int)ThingDefs.Sin:
-                //                    NewTemplate = SinItem;
-                //                    NewTemplateX = SinItemX;
-                //                    break;
-                //                case (int)ThingDefs.Nahkampfwaffe:
-                //                    NewTemplate = NahkampfwaffeItem;
-                //                    NewTemplateX = NahkampfwaffeItemX;
-                //                    break;
-                //                case (int)ThingDefs.Fernkampfwaffe:
-                //                    NewTemplate = FernkampfwaffeItem;
-                //                    NewTemplateX = FernkampfwaffeItemX;
-                //                    break;
-                //                case (int)ThingDefs.Kommlink:
-                //                    NewTemplate = KommlinkItem;
-                //                    NewTemplateX = KommlinkItemX;
-                //                    break;
-                //                case (int)ThingDefs.CyberDeck:
-                //                    NewTemplate = CyberDeckItem;
-                //                    NewTemplateX = CyberDeckItemX;
-                //                    break;
-                //                case (int)ThingDefs.Vehikel:
-                //                    NewTemplate = VehikelItem;
-                //                    NewTemplateX = VehikelItemX;
-                //                    break;
-                //                case (int)ThingDefs.Panzerung:
-                //                    NewTemplate = PanzerungItem;
-                //                    NewTemplateX = PanzerungItemX;
-                //                    break;
-                //                case (int)ThingDefs.Adeptenkraft_KomplexeForm:
-                //                    NewTemplate = Adeptenkraft_KomplexeFormItem;
-                //                    NewTemplateX = Adeptenkraft_KomplexeFormItemX;
-                //                    break;
-                //                case (int)ThingDefs.Geist_Sprite:
-                //                    NewTemplate = Geist_SpriteItem;
-                //                    NewTemplateX = Geist_SpriteItemX;
-                //                    break;
-                //                case (int)ThingDefs.Foki_Widgets:
-                //                    NewTemplate = Foki_WidgetsItem;
-                //                    NewTemplateX = Foki_WidgetsItemX;
-                //                    break;
-                //                case (int)ThingDefs.Stroemung_Wandlung:
-                //                    NewTemplate = Stroemung_WandlungItem;
-                //                    NewTemplateX = Stroemung_WandlungItemX;
-                //                    break;
-                //                case (int)ThingDefs.Tradition_Initiation:
-                //                    NewTemplate = Tradition_InitiationItem;
-                //                    NewTemplateX = Tradition_InitiationItemX;
-                //                    break;
-                //                case (int)ThingDefs.Zaubersprueche:
-                //                    NewTemplate = ZauberspruecheItem;
-                //                    NewTemplateX = ZauberspruecheItemX;
-                //                    break;
-                //                case (int)ThingDefs.KomplexeForm:
-                //                    NewTemplate = KomplexeFormItem;
-                //                    NewTemplateX = KomplexeFormItemX;
-                //                    break;
-                //                case (int)ThingDefs.Sprite:
-                //                    NewTemplate = SpriteItem;
-                //                    NewTemplateX = SpriteItemX;
-                //                    break;
-                //                case (int)ThingDefs.Widgets:
-                //                    NewTemplate = WidgetsItem;
-                //                    NewTemplateX = WidgetsItemX;
-                //                    break;
-                //                case (int)ThingDefs.Wandlung:
-                //                    NewTemplate = WandlungItem;
-                //                    NewTemplateX = WandlungItemX;
-                //                    break;
-                //                case (int)ThingDefs.Initiation:
-                //                    NewTemplate = InitiationItem;
-                //                    NewTemplateX = InitiationItemX;
-                //                    break;
-
-                default:
-                    return;
+                if (!SettingsModel.I.TutorialHandlungShown)
+                {
+#pragma warning disable CS4014
+                    new Tutorial(30, 31).ShowAsync();
+#pragma warning restore CS4014
+                    SettingsModel.I.TutorialHandlungShown = true;
+                }
             }
-
         }
-        private void Grid_Entry_Loaded(object sender, RoutedEventArgs e)
+
+        internal void DefaultTemplate()
         {
+            EntryItem.ContentTemplate = Default;
 
         }
 
-        private void Del_Click(object sender, RoutedEventArgs e)
+        async void Edit_Click(object sender, RoutedEventArgs e)
         {
-
+            try
+            {
+                await new EditThingDialog(((Thing)((Button)sender).DataContext)).ShowAsync();
+            }
+            catch (Exception)
+            {
+            }
         }
 
-        private void Edit_Click(object sender, RoutedEventArgs e)
+        void Del_Click(object sender, RoutedEventArgs e)
         {
+            if ((Thing)((Button)sender).DataContext != null)
+            {
+                Model.MainObject.Remove((Thing)((Button)sender).DataContext);
+            }
+        }
+
+        async void HandlungEditZusDialog_Click(object sender, RoutedEventArgs e)
+        {
+            Auswahl dialog = new Auswahl(Model.MainObject.LinkList, ((Handlung)((Button)sender).DataContext).WertZusammensetzung, Filter: Handlung.Filter);
+            await dialog.ShowAsync();
 
         }
 
+        async void HandlungEditGrenzeZusDialog_Click(object sender, RoutedEventArgs e)
+        {
+            Auswahl dialog = new Auswahl(Model.MainObject.LinkList, ((Handlung)((Button)sender).DataContext).GrenzeZusammensetzung, Filter: Handlung.Filter);
+            var ergebnis = await dialog.ShowAsync();
+        }
+
+        async void HandlungEditGegenZusDialog_Click(object sender, RoutedEventArgs e)
+        {
+            Auswahl dialog = new Auswahl(Model.MainObject.LinkList, ((Handlung)((Button)sender).DataContext).GegenZusammensetzung, Filter: Handlung.Filter);
+            await dialog.ShowAsync();
+        }
+
+        async void FertigkeitenZusammensetzungBearbeiten(object sender, RoutedEventArgs e)
+        {
+            Auswahl dialog = new Auswahl(Model.MainObject.LinkList, ((Fertigkeit)((Button)sender).DataContext).PoolZusammensetzung, Filter: Fertigkeit.Filter);
+            await dialog.ShowAsync();
+        }
+
+        async void Edit_LinkedThings(object sender, RoutedEventArgs e)
+        {
+            ThingDefs type = (ThingDefs)int.Parse((sender as FrameworkElement).Tag as string);
+            IEnumerable<ThingDefs> FilterToUse = null;
+            switch (type)
+            {
+                case ThingDefs.Handlung:
+                    FilterToUse = Handlung.Filter;
+                    break;
+                case ThingDefs.Fertigkeit:
+                    FilterToUse = Fertigkeit.Filter;
+                    break;
+                case ThingDefs.Item:
+                    break;
+                case ThingDefs.Programm:
+                    break;
+                case ThingDefs.Munition:
+                    break;
+                case ThingDefs.Implantat:
+                    break;
+                case ThingDefs.Vorteil:
+                    break;
+                case ThingDefs.Nachteil:
+                    break;
+                case ThingDefs.Attribut:
+                    FilterToUse = Attribut.Filter;
+                    break;
+                case ThingDefs.Nahkampfwaffe:
+                    break;
+                case ThingDefs.Fernkampfwaffe:
+                    FilterToUse = Fernkampfwaffe.Filter;
+                    break;
+                case ThingDefs.Kommlink:
+                    break;
+                case ThingDefs.CyberDeck:
+                    break;
+                case ThingDefs.Vehikel:
+                    break;
+                case ThingDefs.Panzerung:
+                    FilterToUse = Panzerung.Filter;
+                    break;
+                case ThingDefs.Eigenschaft:
+                    break;
+                case ThingDefs.Adeptenkraft_KomplexeForm:
+                    break;
+                case ThingDefs.Geist_Sprite:
+                    break;
+                case ThingDefs.Foki_Widgets:
+                    break;
+                case ThingDefs.Stroemung_Wandlung:
+                    break;
+                case ThingDefs.Tradition_Initiation:
+                    break;
+                case ThingDefs.Zaubersprueche:
+                    break;
+                case ThingDefs.Berechnet:
+                    break;
+                case ThingDefs.KomplexeForm:
+                    break;
+                case ThingDefs.Sprite:
+                    break;
+                case ThingDefs.Widgets:
+                    break;
+                case ThingDefs.Wandlung:
+                    break;
+                case ThingDefs.Initiation:
+                    break;
+                default:
+                    break;
+            }
+            var dialog = new Auswahl(Model.MainObject.LinkList, ((Thing)((Button)sender).DataContext).LinkedThings, Filter: FilterToUse);
+            await dialog.ShowAsync();
+        }
+        //async void MunitionBearbeiten(object sender, RoutedEventArgs e)
+        //{
+        //    var TemList = new ObservableCollection<AllListEntry> ();
+        //    Auswahl dialog = new Auswahl(Model.MainObject.LinkList, TemList, false, Fernkampfwaffe.Filter);
+        //    await dialog.ShowAsync();
+        //    ((Fernkampfwaffe)((Button)sender).DataContext).CurrentMunition = TemList.FirstOrDefault();
+        //}
     }
 }
